Rewrite validate imports as an analyzer (#156)
Validate imports was failing after the repo migration, because it was using $GOPATH to open and read files. Fix by writing it as an analyzer and running via roxvet -- the check logic is pretty much the same, but we are now able to push all the logic of walking the directory/parsing files to the analyzer framework.

Also, move all roxvet analyzers into a directory under roxvet, because the tools directory has become unwieldy.