Reset ClusterMetrics on disconnect (#1952)
The current implementation does not explicitly reset the cluster metrics
upon a disconnect of the Sensor <-> Central connection. This can be an
issue because the reported cluster metrics Central could be outdated
without being marked as stale by Prometheus.

The latter occurs only when Central itself is offline. However, if
Sensor is offline, Central keeps the last known gauge value in memory
and reports it to Prometheus. The solution is an explicit reset of the
cluster metrics. With the proposed implementation, the cluster metrics
will be reset to their zero value if the connection drops. They will
pick up again once a new connection is established.