Fix handling empty string matches (#5211)
Trimming values and building new string query do not handle empty
strings correctly.

The first one is missing out to extract it out of double quotes due to
the length condition (an empty string is `""`, which is exactly the same
lenght as 2 * `"`, where the `"` is EqualityPrefixSuffix). This leads to
an empty string being represented as `\"\"` along the way, producing
incorrect search results.

The second one assumes that an empty string is something incorrect,
although it's totally fine to search for values equal empty string.