ROX-18391: cleanup permission checker
## Description

When we started PG migration project we were unsure in what direction it will go with permission checker. This resulted in a safe approach to support everything we could think of. After time we can cleanup permission checker interface and limit it to only two methods.

## Checklist
- [ ] Investigated and inspected CI test results
- [ ] Unit test and regression tests added
- [ ] Evaluated and added CHANGELOG entry if required
- [ ] Determined and documented upgrade steps
- [ ] Documented user facing changes (create PR based on [openshift/openshift-docs](https://github.com/openshift/openshift-docs) and merge into [rhacs-docs](https://github.com/openshift/openshift-docs/tree/rhacs-docs))

If any of these don't apply, please comment below.

## Testing Performed

CI