refactor(pg): store nil map as empty {} (#6929)
## Description

Currently we store nil map as `'null'` (string). In next version of pgx the semantic has changed and it will be stored as `null`. 
This PR seal this behaviour and whenever we would like to store nil map it will be replaced with `{}` (empty but not nil map).

`TestIndex/TestMapHighlights/key_does_not_exist` shows the difference between having `'null'` and `null` in JSON field and null always returns false so negative queries will not work correctly.  

## Checklist
- [ ] Investigated and inspected CI test results
- [ ] Unit test and regression tests added
- [ ] Evaluated and added CHANGELOG entry if required
- [ ] Determined and documented upgrade steps
- [ ] Documented user facing changes (create PR based on [openshift/openshift-docs](https://github.com/openshift/openshift-docs) and merge into [rhacs-docs](https://github.com/openshift/openshift-docs/tree/rhacs-docs))

If any of these don't apply, please comment below.

## Testing Performed

CI