ROX-18313: Fix compliance integration test runner
Currently, we have e2e testing that exercises the Compliance Operator
integration, which is invoked with all e2e/integration tests.

But, there is also a compliance-specific e2e suite that only sets up and
runs compliance-related integration tests using:

  make -C tests compliance-tests

Right now, this suite doesn't run any tests because the go test
invocation uses `-run compliance_operator_test.go`, when it's expecting
a regular expression:

  The argument to the -run, -bench, and -fuzz command-line flags is an
  unanchored regular expression that matches the test's name.

This commit uses a consistent substring in all the compliance-related
tests, and then using that string as a regular expression in go test
invocation, so they are discovered when you use the `compliance-tests`
target.

This isn't a problem when we run all e2e tests using `run.sh` because we
do test discovery without using `-run`, or a regular expression, so all
compliance-related tests are discovered.

This will become more useful when we have compliance tests running with
a dedicated e2e test that gets invoked on all compliance-related
changes.