ROX-22144: Disallow creating a scan configuration if SSB has same profile
With compliance operator, if a profile is executed under multiple scan setting bindings, the checks only run once.  So to ensure that the user retrieves results as expected then we should alert the user if the scan configuration being created in ACS clashes with an existing SSB in compliance operator.  If such a clash exists we should return an error notifying the user of this during creation of the scan configuration.