ROX-25484: Install k8s service account m2m auth (#12237)
In order to guarantee that the config-controller (the component that will reconcile k8s resources against the Central API) will be able to authenticate against Central out-of-the-box, the m2m configurtion will be hard-coded in to Central such that it will be installed during bootstrap.

The configuration for the m2m auth is as follows:

* The issuer URL is pulled from the service account token available to the Central pod
* Only the config-controller service account will be granted any roles with this particular m2m config
* Of type `KUBE_SERVICE_ACCOUNT`

There is a new m2m type because by default K8s requires authentication for fetching the OIDC resources required to validate service account tokens.
See: https://kubernetes.io/docs/tasks/configure-pod-container/configure-service-account/#service-account-issuer-discovery

That said, some K8s distros (namely GCP) actually fail requests for OIDC configuration when a k8s service account token provided in the auth header, so the request must first be tried without the token to see if that succeeds.  If not, then the token is added.

This commit also adds a new ACS role called "Configuration Controller" that grants the config-controller its required access (and no more).