ROX-25490: Populate SecurityPolicy CRD spec; caching gRPC client with m2m auth (#12647)
Co-authored-by: Khushboo Sancheti <ks@stackrox.com>

Resolves ROX-26052
Resolves ROX-25729

## Caching layer in client

Caching policies in the controller so that it doesn't put too much of a load on Central. The cache will be updated on the first reconciliation that happens after the cache age hits five minutes.

Note that the cache is a map, keyed off the policy name. This means the reconciler code could be simplified a bit to avoid looping over all policies looking for a match on name.

I also refactored the CR status update code a bit to hopefully make it clearer.

## Client auth and use of public client APIs

### Problem

The grpc client needed to be reworked for a couple reasons:

* In my PoC I just used the first code I could find (roxctl). These APIs were actually private, so I needed to use "public" APIs under pkg.
* It needed to use the m2m endpoint to periodically exchange the Kube service account token for a short-lived Central API token.

### Details of implementation

* Token refresher uses same grpc connection as policy client
* The token is only refreshed on reconciles, similar to cache
* Added `TokenExchange` method to `PolicyClient` interface, which gets invoked via `CachedPolicyClient.EnsureFresh()`.
* Set grpc user agent to "Config Controller"
* Wired up the context for the client to cancel from signals
* Updated Kube service account m2m auth config to expire tokens every 1m
* In order to trust Central's cert, the CA from `central-tls` secret is mounted into config-controller container.
* gRPC connection opts:
    1. No proxy
    2. Max message size of 12MB
    3. Don't allow insecure TLS

### Other Changes

A couple of other changes included:

* Configure logging for client.  Added a log message for every call to Central.  Maybe too verbose for a release, but it will probably be useful for now at least.
* Make the namespace in Central's hostname dynamic by using POD_NAMESPACE env var.
* I didn't realize returning an error from Reconcile() would auto-requeue, so changed the code accordingly.

## Description of CRD definition changes

This is a proposed way to keep the policy protobuf and CRD sepc in sync.  I wrote out the CRD spec structs by hand as well as a `ToProtobuf()` function that converts the CRD spec to a `storage.Policy` protobuf struct.

To test:

* Read in a policy CRD spec JSON file and unmarshal into the CRD structs
* Validate it is read as expected
* Convert it to `storage.Policy`
* Marshal that into JSON
* Compare the resulting JSON (with a bit of massaging) to the original JSON