chore(load-generators): Reduce process load (#13308)
Currently the process spawning rate is 10, which is roughly 10 processes per second per worker. Normally we got 8 cores on release cluster nodes, meaning that it produces 80 processes/s for a single Collector. If there are less nodes that berserkers, some of them land on the same machine making load even higher.

This leads to situations, when the load is just too high to get any meaningful results, making the whole setup unstable and imbalanced. As a workaround for now, reduce spawning rate, to produce milder load.