ROX-20769: Add backup PVC (#14704)
For PG major version upgrade we need to have a storage to keep a backup. Since
PVCs are not managed by Helm, implement that directly in the reconciling logic.

The backup PVC will use the same configuration as the main PVC, with a few
exceptions:

* The name will be derived from the main PVC with a suffix "-backup". Thus a
default name will be "central-db-backup", and in case if it was overwritten to
"some-other-name", the backup volume will be "some-other-name-backup".

* The size will be derived from the main PVC as double if its size. The idea is
that the backup volume has to accommodate at least one backup and one restored
database (when we test a freshly taken backup). In the worst case scenario
the backup and the restored db will take the same amount of space, hence the
doubling.

Keep in mind, that the logic for volume reconciliation is changed in this PR.
Due to concerns about setups where there is no default storage class
configured, we now verify if that's the case before creating a volume. The
logic is to skip creation if no default storage class is present and no
explicit storage class is defined in the manifest. This should be relevant only
for the backup volumes, as the main data volume should be already created after
upgrading to the new operator version.

Co-authored-by: Marcin Owsiany <porridge@redhat.com>